Refactor custom middlewares out of server
