+++ src/errors/validationError.ts
@@ -1,0 +1,3 @@
+class ValidationError extends Error {}
+
+export { ValidationError };
