+++ src/middleware/index.ts
@@ -1,0 +1,2 @@
+export { errorHandling } from './errorHandling';
+export { logging } from './logging';
